issue#4826 Impedido que seja excluído o usuário Administrador ou o próprio usuário

+++ backend/projeto/escoladeti/src/main/java/br/unicesumar/escoladeti/service/UsuarioService.java
@@ -1,5 +1,6 @@
 package br.unicesumar.escoladeti.service;
 
+import br.unicesumar.escoladeti.application.Sessao;
 import br.unicesumar.escoladeti.controller.DataPage;
 import static br.unicesumar.escoladeti.controller.DataPage.pageRequestForAsc;
 import br.unicesumar.escoladeti.dto.UsuarioDTO;
@@ -12,6 +13,7 @@ import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -22,10 +24,10 @@ public class UsuarioService {
     private static final Logger logger = LoggerFactory.getLogger(UsuarioService.class);
     @Autowired
     private UsuarioRepository usuarioRepository;
-    
+
     @Autowired
     private PerfilAcessoUsuarioRepository perfilAcessoUsuarioRepository;
-    
+
     @Autowired
     private PerfilDeAcessoUsuarioViewRepository perfilDeAcessoUsuarioViewRepository;
 
@@ -55,7 +57,7 @@ public class UsuarioService {
             perfilAcesso.setNome("ADMINISTRADOR");
             List<ItemAcesso> todosItens = itemAcessoRepository.findAll();
 
-            for(ItemAcesso itemAcesso: todosItens) {
+            for (ItemAcesso itemAcesso : todosItens) {
                 perfilAcesso.adcionarItem(itemAcesso.getId());
             }
 
@@ -76,62 +78,85 @@ public class UsuarioService {
         logger.info("Usuário 'admin' verificado.");
     }
 
-
-
     public UsuarioDTO salvar(UsuarioDTO usuarioDTO) {
-        Usuario usuario = new Usuario();
-        usuario.setId(usuarioDTO.getId());
-        usuario.setLogin(usuarioDTO.getLogin());
-        usuario.setNome(usuarioDTO.getNome());
-        usuario.setSenha(usuarioDTO.getSenha());
-        usuario.setEmail(usuarioDTO.getEmail());
-        usuario.setAtivo(usuarioDTO.getAtivo());
         
-        Usuario usuSalvo = this.usuarioRepository.save(usuario);
-         
-        PerfilAcessoUsuario usuarioperfilacesso = new PerfilAcessoUsuario();
-        usuarioperfilacesso.setId(usuarioDTO.getPerfilDeAcessoUsuarioId());
-        usuarioperfilacesso.setInicioVigencia(usuarioDTO.getInicioVigencia());
-        usuarioperfilacesso.setFimVigencia(usuarioDTO.getFimVigencia());
-        usuarioperfilacesso.setUsuario(usuSalvo);
+        Usuario usuarioConsulta = getUsuarioRepository().findByLogin(usuarioDTO.getLogin());
         
-        PerfilAcesso pa = new PerfilAcesso();
-        pa.setId(usuarioDTO.getPerfilAcessoId());
-        pa.setNome("");
-        usuarioperfilacesso.setPerfilAcesso(pa);
+        if (usuarioConsulta == null) {
         
-        this.perfilAcessoUsuarioRepository.save(usuarioperfilacesso);
+            Usuario usuario = new Usuario();
+
+            usuario.setId(usuarioDTO.getId());
+            usuario.setLogin(usuarioDTO.getLogin());
+            usuario.setNome(usuarioDTO.getNome());
+            usuario.setSenha(usuarioDTO.getSenha());
+            usuario.setEmail(usuarioDTO.getEmail());
+            usuario.setAtivo(usuarioDTO.getAtivo());
+
+            Usuario usuSalvo = this.usuarioRepository.save(usuario);
+
+            PerfilAcessoUsuario usuarioperfilacesso = new PerfilAcessoUsuario();
+
+            usuarioperfilacesso.setId(usuarioDTO.getPerfilDeAcessoUsuarioId());
+            usuarioperfilacesso.setInicioVigencia(usuarioDTO.getInicioVigencia());
+            usuarioperfilacesso.setFimVigencia(usuarioDTO.getFimVigencia());
+            usuarioperfilacesso.setUsuario(usuSalvo);
+
+            PerfilAcesso pa = new PerfilAcesso();
+
+            pa.setId(usuarioDTO.getPerfilAcessoId());
+            pa.setNome("");
+
+            usuarioperfilacesso.setPerfilAcesso(pa);
+
+            this.perfilAcessoUsuarioRepository.save(usuarioperfilacesso);
+
+        } else {
+            throw new RuntimeException("Login não pode ser repetido.");
+        }
         
         return usuarioDTO;
     }
-    
+
     public DataPage<ViewPerfilDeAcessoUsuario> listarUsuarios(Integer pagina) {
-       return new DataPage<>(perfilDeAcessoUsuarioViewRepository.findAll(pageRequestForAsc(pagina, "nome")));
+        return new DataPage<>(perfilDeAcessoUsuarioViewRepository.findAll(pageRequestForAsc(pagina, "nome")));
     }
-    
+
     public DataPage<ViewPerfilDeAcessoUsuario> getPerfilDeAcessoUsuarioPorNomeOrPerfilOrEmail(String nomeParcial) {
-        return new DataPage<ViewPerfilDeAcessoUsuario>
-        (perfilDeAcessoUsuarioViewRepository.findByNomeContainingOrPerfilContainingOrEmailContainingOrderByNomeAsc
-        (nomeParcial, nomeParcial, nomeParcial, pageRequestForAsc(1, "nome")));  
+        return new DataPage<ViewPerfilDeAcessoUsuario>(perfilDeAcessoUsuarioViewRepository.findByNomeContainingOrPerfilContainingOrEmailContainingOrderByNomeAsc(nomeParcial, nomeParcial, nomeParcial, pageRequestForAsc(1, "nome")));
     }
-    
+
     public ViewPerfilDeAcessoUsuario getById(Long id) {
         return perfilDeAcessoUsuarioViewRepository.findById(id);
     }
-    
+
     public void deletar(UsuarioDTO usuarioDTO) {
-        Usuario usuario = new Usuario();
-        usuario.setId(usuarioDTO.getId());
+        PerfilAcessoUsuario perfilAcessoUsuario
+                = perfilAcessoUsuarioRepository.findUniqueByUsuarioId(usuarioDTO.getId());
 
-        PerfilAcessoUsuario perfilAcessoUsuario = new PerfilAcessoUsuario();
-        perfilAcessoUsuario.setUsuario(usuario);
+        if (!perfilAcessoUsuario.getPerfilAcesso().getNome().equalsIgnoreCase("ADMINISTRADOR")) {
+            
+            UserDetails userDetails = Sessao.currentUserDetails();
+            if (!userDetails.getUsername().equalsIgnoreCase(usuarioDTO.getLogin())) {
 
-        this.perfilAcessoUsuarioRepository.deleteByUsuario(usuario);
-        this.usuarioRepository.delete(usuario);
+                Usuario usuario = new Usuario();
+                usuario.setId(usuarioDTO.getId());
+
+//                PerfilAcessoUsuario perfilAcessoUsuario = new PerfilAcessoUsuario();
+                perfilAcessoUsuario.setUsuario(usuario);
+
+                this.perfilAcessoUsuarioRepository.deleteByUsuario(usuario);
+                this.usuarioRepository.delete(usuario);
+            } else {
+                throw new RuntimeException("O usuário não pode se exluir");
+            }
+
+        } else {
+            throw new RuntimeException("Impossível exluir o usuário administrador do sistema");
+        }
     }
-    
-    public List<Usuario> getTodos(){
+
+    public List<Usuario> getTodos() {
         return usuarioRepository.findAll();
     }
-   
 }
